requirements.txt
discord.py
discord-py-slash-command
requests
cfscrape
lxml
aiohttp
tweepy
py-expression-eval
beautifulsoup4
pytz
beautifultable==1.0.0
psutil
fuzzywuzzy
dblpy
Pillow
imagehash